rubocopで修正 #41
