+++ app/javascript/stylesheets/rate.scss
@@ -16,7 +16,6 @@
 .rate-container {
   width: 60%;
   max-width: 700px;
-  // border-radius: 10px;
   margin: 0rem auto;
   padding: 1.4rem 3rem;
   border: solid 1px $color-trueBlack;
@@ -49,9 +48,6 @@
       }
     }
   }
-}
-
-.rate-bottom {
 }
 
 .rate-currency {
